v2.4.3: Performance and Compatibility Improvements

+++ miner.h
@@ -152,12 +152,12 @@ void sha256_transform_8way(uint32_t *state, const uint32_t *block, int swap);
 #endif
 
 extern int scanhash_sha256d(int thr_id, uint32_t *pdata,
-	const uint32_t *ptarget, uint32_t max_nonce, unsigned long *hashes_done);
+  const uint32_t *ptarget, uint32_t max_nonce, unsigned int *hashes_done);
 
 extern unsigned char *scrypt_buffer_alloc();
 extern int scanhash_scrypt(int thr_id, uint32_t *pdata,
-	unsigned char *scratchbuf, const uint32_t *ptarget,
-	uint32_t max_nonce, unsigned long *hashes_done);
+  unsigned char *scratchbuf, const uint32_t *ptarget,
+  uint32_t max_nonce, unsigned int *hashes_done);
 
 struct thr_info {
 	int		id;
