+++ sha2.c
@@ -469,8 +469,7 @@ void sha256d_ms_4way(uint32_t *hash,  uint32_t *data,
 	const uint32_t *midstate, const uint32_t *prehash);
 
 static inline int scanhash_sha256d_4way(int thr_id, uint32_t *pdata,
-	const uint32_t *ptarget, uint32_t max_nonce, unsigned long *hashes_done)
-{
+  const uint32_t *ptarget, uint32_t max_nonce, unsigned int *hashes_done) {
 	uint32_t data[4 * 64] __attribute__((aligned(128)));
 	uint32_t hash[4 * 8] __attribute__((aligned(32)));
 	uint32_t midstate[4 * 8] __attribute__((aligned(32)));
@@ -528,8 +527,7 @@ void sha256d_ms_8way(uint32_t *hash,  uint32_t *data,
 	const uint32_t *midstate, const uint32_t *prehash);
 
 static inline int scanhash_sha256d_8way(int thr_id, uint32_t *pdata,
-	const uint32_t *ptarget, uint32_t max_nonce, unsigned long *hashes_done)
-{
+  const uint32_t *ptarget, uint32_t max_nonce, unsigned int *hashes_done) {
 	uint32_t data[8 * 64] __attribute__((aligned(128)));
 	uint32_t hash[8 * 8] __attribute__((aligned(32)));
 	uint32_t midstate[8 * 8] __attribute__((aligned(32)));
@@ -582,8 +580,7 @@ static inline int scanhash_sha256d_8way(int thr_id, uint32_t *pdata,
 #endif /* HAVE_SHA256_8WAY */
 
 int scanhash_sha256d(int thr_id, uint32_t *pdata, const uint32_t *ptarget,
-	uint32_t max_nonce, unsigned long *hashes_done)
-{
+  uint32_t max_nonce, unsigned int *hashes_done) {
 	uint32_t data[64] __attribute__((aligned(128)));
 	uint32_t hash[8] __attribute__((aligned(32)));
 	uint32_t midstate[8] __attribute__((aligned(32)));
